cegrisa/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
python-dotenv==1.1.1
sqlparse==0.5.3
